Make ArbReplayableTx and ArbStatistics only compile in the test configuration
